% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyItemLinks.R
\name{etsyItemLinks}
\alias{etsyItemLinks}
\title{Function to obtain URL links for each item returned by an Etsy.com search.}
\usage{
etsyItemLinks(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage.}
}
\value{
Returns a character vector of URLs, containing the unique URL for each item resulting from an Etsy.com search.
}
\description{
This function returns the unique URL for each item resulting from an Etsy.com search.
}
\examples{
etsyItemLinks(results_xml)
etsyItemLinks(xml_code)

}
\keyword{html}
\keyword{item}
\keyword{xml}
