% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyItemResults.R
\name{etsyItemResults}
\alias{etsyItemResults}
\title{Function to obtain the names of the items returned from an Etsy.com search.}
\usage{
etsyItemResults(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage.

\emph{Note: The xml is easily obtained using the etsyURL() and etsyXML() functions.}}
}
\value{
Returns the names of items that result from an Etsy.com search as a character vector.
}
\description{
This function obtains the names of the items returned from an Etsy.com search by reading
the scraped xml.
}
\examples{
etsyItemResults(results_xml)
etsyItemResults(xml_code)

}
\keyword{html}
\keyword{item}
\keyword{xml}
