% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyscrape.R
\docType{package}
\name{etsyscrape}
\alias{etsyscrape}
\alias{etsyscrape-package}
\title{etsyscrape: Etsy Web Scraping Tools}
\description{
This packages allows the user to examine search results scraped from Etsy.com.
The package contains functions to build an Etsy URL based on a search string and desired ordering
of results, and functions to gather information about the names of the items, the shops they are sold
in, the location of the shop, and links to the items themselves.
}
\section{etsyscrape functions}{

\itemize{
  \item etsyURL
  \item etsyXML
  \item etsyItemResults
  \item etsyItemLink
  \item etsyShopResults
  \item etsyShopLocation
}
}

\author{
Meagan McCardle \email{meagan_mccardle@brown.edu}

Nana Boame \email{nana_boame@brown.edu}
}
