% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyShopLocations.R
\name{etsyShopLocations}
\alias{etsyShopLocations}
\title{Function to obtain shop location.}
\usage{
etsyShopLocations(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage.}
}
\value{
Returns the shop location as a character vector.
}
\description{
etsyShopLocations scrapes the location of each etsy shop as listed on the shop's webpage
The function uses the etsyGetShops function to extract the names of the relevent etsy shops.
It then creates a url for each individual shop, and scrapes that URL for the shop location. 
The function returns the shop locations in a vector.
}
\keyword{html}
\keyword{location}
\keyword{xml}
