% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyShopResults.R
\name{etsyShopResults}
\alias{etsyShopResults}
\title{Function to obtain the shop names of the shops selling the items returned by an Etsy.com search.}
\usage{
etsyShopResults(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage.}
}
\value{
Returns the shop names that result from an Etsy.com search as a character vector.
}
\description{
This function obtains the names of the shops returned from an Etsy.com search by reading
the scraped xml.
}
\keyword{html}
\keyword{shops}
\keyword{xml}
