% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyShopReviews.R
\name{etsyShopReviews}
\alias{etsyShopReviews}
\title{Function to obtain shop reviews.}
\usage{
etsyShopReviews(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage}
}
\value{
Returns the shop reviews as a character vector.
}
\description{
This function scrapes the text-based customer reviews of each shop resulting from the etsy.com search
The function uses the etsyGetShops function to extract the names of the relevent etsy shops.
It then creates a url for each individual shop, and scrapes that URL for the shop reviews.
The function collapses all the different reviews on the first page (only) of the shop's review url into one text paragraph,
and returns a vector of paragraphs representing the reviews from each shop.
}
\keyword{html}
\keyword{reviews}
\keyword{xml}
