% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyURL.R
\name{etsyURL}
\alias{etsyURL}
\title{Function to create Etsy.com URL for item search.}
\usage{
etsyURL(search_string, search_order = "most_relevant")
}
\arguments{
\item{search_string}{What item would you like to search for?}

\item{search_order}{Specify the order in which the search results should be returned.
Default is "most_relevent".
Other options include "date_desc", "price_desc", "price_asc".}
}
\value{
Returns an Etsy.com URL as a string.
}
\description{
This function creates the URL associated with an Etsy.com search for a specific item.
}
\examples{
etsyURL("knitted scarf")
etsyURL("welcome sign", "date_desc")

}
\keyword{url}
