% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyXML.R
\name{etsyXML}
\alias{etsyXML}
\title{Function to read and return the html of an Etsy search URL.}
\usage{
etsyXML(url)
}
\arguments{
\item{url}{The URL of the Etsy.com search.}
}
\value{
Returns the xml from scraping the provided URL.
}
\description{
This function accepts the URL of an Etsy.com search and scrapes the html.
}
\examples{
etsyXML("https://www.etsy.com/search?q=hello+kitty+purse&explicit=1&order=most_relevant")

}
\keyword{html}
\keyword{url}
\keyword{xml}
