% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsyShopLocations.R
\name{etsyShopLocations}
\alias{etsyShopLocations}
\title{Function to obtain shop locations of the shops selling the items returned by an Etsy.com search.}
\usage{
etsyShopLocations(results_xml)
}
\arguments{
\item{results_xml}{An xml object, obtained from scraping an Etsy search results webpage.}
}
\value{
Returns the shop locations for all shops selling an item returned by the search results,
as a character vector.
}
\description{
This function scrapes the location of each Etsy shop as listed on the shop's homepage.
\itemize{
  \item Utilizes the etsyShopResults() function to extract the name of each Etsy shop from the
  Etsy.com search results.
  \item Creates a URL for each individual shop, and scrapes the shop URL for the shop location.
}
}
\keyword{html}
\keyword{location}
\keyword{xml}
